whu-city-backend/requirements.txt
fastapi>=0.104.1,<0.116
uvicorn>=0.25.0
pydantic>=2.7.4,<3.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
dashscope>=1.17.0

